Peters-Belson Disparity Analysis

Function pb offers Peters-Belson(PB) type of regression method which gets the disparity between a majority group and a minority group based on various regression models.

pb(formula, group, data, family = "gaussian")

Arguments

formula

a formula for the model.

group

a vector within the data which is used for separating majority and minority groups.

data

a data frame and data has to be included with the form of data.frame.

family

a character indicating which model should be used. Details can be found later.

Value

pb returns an object of class "pb", which is a list containing following components:

call

a matched call.

overall_disparity

overall disparity between major and minor groups.

explained_disparity

explained disparity between major and minor groups.

unexplained_disparity

unexplained disparity between major and minor groups.

major

a majority group label.

minor

a minority group label.